Refactor: Move activity demo code to another activity
