+++ app/src/main/AndroidManifest.xml
@@ -21,5 +21,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <activity android:name=".VGSShowActivity" />
     </application>
 </manifest>
