+++ app/src/main/res/drawable/ic_vgs_logo_grey_64dp.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="82dp"
+    android:height="92dp"
+    android:viewportWidth="82"
+    android:viewportHeight="92">
+
+    <path
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:pathData="M41.207 91.6708L41.2777 91.6312L41.3486 91.6708L61.4043 80.3451L81.611 68.9404V23.3143L72.525 28.4434L67.2216 31.7173L55.6355 38.3295V48.7664L72.525 39.1725V55.7811L72.5243 55.7816L72.3869 63.6333L41.2777 81.1954L10.1687 63.6333L10.031 55.7816L10.0306 55.7811V28.4434L10.1008 28.4039L10.1687 28.4434L41.2777 10.8827L63.2125 23.2647L72.525 18.0077L41.3468 0.407715L41.2777 0.44687L41.2087 0.407715L0.944336 23.137V68.9404L21.151 80.3451L41.207 91.6708Z" />
+    <path
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:pathData="M16.8333 32.0459L26.8863 37.7137V52.5949L41.2775 60.6922L55.669 52.5949V49.9484L65.7221 44.2914V58.4475L41.2775 72.2016L16.8333 58.4475V32.0459Z" />
+</vector>
